refactoring and improving biocrates2isa.py, adding tests, adding resources
