+++ tests/output/isatab/i_inv_biocrates.txt
@@ -1,0 +1,91 @@
+"ONTOLOGY SOURCE REFERENCE"
+"Term Source Name"	"OBI"	"PSI-MS"	"UBERON"	"NCBITax"
+"Term Source File"	"http://purl.org/obo/obi.owl"	"http://psi-ms.sf.net"	"http://purl.org/obo/uberon.owl"	"http://purl.org/obo/NCBITax.owl"
+"Term Source Version"	"1"	"1"	"1"	"1"
+"Term Source Description"	"The Ontology for Biomedical Investigation"	"PSI Mass Spectrometry"	"Uber-anatomy ontology"	"NCBI Taxonomy"
+"INVESTIGATION"
+"Investigation Identifier"
+"Investigation Title"
+"Investigation Description"
+"Investigation Submission Date"
+"Investigation Public Release Date"
+"INVESTIGATION PUBLICATIONS"
+"Investigation PubMed ID"
+"Investigation Publication DOI"
+"Investigation Publication Author List"
+"Investigation Publication Title"
+"Investigation Publication Status"
+"Investigation Publication Status Term Accession Number"
+"Investigation Publication Status Term Source REF"
+"INVESTIGATION CONTACTS"
+"Investigation Person Last Name"
+"Investigation Person First Name"
+"Investigation Person Mid Initials"
+"Investigation Person Email"
+"Investigation Person Phone"
+"Investigation Person Fax"
+"Investigation Person Address"
+"Investigation Person Affiliation"
+"Investigation Person Roles"
+"Investigation Person Roles Term Accession Number"
+"Investigation Person Roles Term Source REF"
+"STUDY"
+"Study Identifier"	"MTLBSXXX_78611000000247426927861100000024802295"
+"Study Title"	"Resolute"
+"Study Submission Date"	""
+"Study Public Release Date"	""
+"Study Description"	""
+"Study File Name"	"s_study_biocrates.txt"
+"STUDY DESIGN DESCRIPTORS"
+"Study Design Type"
+"Study Design Type Term Accession Number"
+"Study Design Type Term Source REF"
+"STUDY PUBLICATIONS"
+"Study PubMed ID"
+"Study Publication DOI"
+"Study Publication Author List"
+"Study Publication Title"
+"Study Publication Status"
+"Study Publication Status Term Accession Number"
+"Study Publication Status Term Source REF"
+"STUDY FACTORS"
+"Study Factor Name"
+"Study Factor Type"
+"Study Factor Type Term Accession Number"
+"Study Factor Type Term Source REF"
+"STUDY ASSAYS"
+"Study Assay Measurement Type"	"metabolite profiling"	
+"Study Assay Measurement Type Term Source REF"	"OBI"	
+"Study Assay Measurement Type Term Accession Number"	"http://purl.obolibrary.org/obo/OBI_0000366"	
+"Study Assay Technology Type"	"mass spectrometry"	
+"Study Assay Technology Type Term Source REF"	"OBI"	
+"Study Assay Technology Type Term Accession Number"	"http://purl.obolibrary.org/obo/OBI_0000470"	
+"Study Assay Technology Platform"	"Biocrates"	
+"Study Assay File Name"	"a_biocrates_assay_positive_mode.txt"	
+"STUDY PROTOCOLS"
+"Study Protocol Name"	"sample collection"	"extraction/derivatization"	"KIT3-0-8015-MS acquisition in positive mode"	"KIT2-0-8015-MS acquisition in positive mode"	"biocrates analysis"	
+"Study Protocol Type"	"specimen collection"	"material separation"	"data collection"	"data collection"	"data transformation"	    
+"Study Protocol Type Term Accession Number"	"http://purl.obolibrary.org/obo/OBI_0000659"	"http://purl.obolibrary.org/obo/OBI_0302884"	"http://purl.obolibrary.org/obo/OBI_0600013"	"http://purl.obolibrary.org/obo/OBI_0600013"	"http://purl.obolibrary.org/obo/OBI_0200000"	        
+"Study Protocol Type Term Source REF"	"OBI"	"OBI"	"OBI"	"OBI"	"OBI"	
+"Study Protocol Description"
+"Study Protocol URI"
+"Study Protocol Version"
+"Study Protocol Parameters Name"			"kit name;plate ID;plate position;mass spectrometry instrument;acquisition parameter file;inlet type;polarity;run number;injection number"	"kit name;plate ID;plate position;mass spectrometry instrument;acquisition parameter file;inlet type;polarity;run number;injection number"	"software"	
+"Study Protocol Parameters Name Term Accession Number"
+"Study Protocol Parameters Name Term Source REF"
+"Study Protocol Components Name"
+"Study Protocol Components Type"
+"Study Protocol Components Type Term Accession Number"
+"Study Protocol Components Type Term Source REF"
+"STUDY CONTACTS"  
+"Study Person Last Name"	"BIOCRATES"	""
+"Study Person First Name"	"Life Sciences AG"	""       
+"Study Person Mid Initials"
+"Study Person Email"	"N/A"	"N/A" 
+"Study Person Phone"	"N/A"	"N/A" 
+"Study Person Fax"
+"Study Person Address"	", , , "	", , , " 
+"Study Person Affiliation"	"BIOCRATES Life Sciences AG"	"CeMM" 
+"Study Person Roles"
+"Study Person Roles Term Accession Number"
+"Study Person Roles Term Source REF"
